scripts improvements + new cleanup.sh scripts

+++ uninstaller.sh
@@ -1,6 +1,16 @@
 #!/bin/bash
-rm -rf ~/.WineApps/Adobe-Photoshop
+
+if [ "$1" = '' ]; then
+    echo "Error: You are missing path where photoshop is installed eg. "/home/user/""
+    exit 1
+fi
+
+rm -rf $1/Adobe-Photoshop
 rm -rf ~/.local/share/icons/photoshop.png
 rm -rf  ~/.local/share/applications/photoshop.desktop
+rm -rf allredist.tar.xz
+rm -rf 'allredist/'
+rm -rf winetricks
+rm -rf 'Adobe Photoshop 2021/' 
 
-zenity --info --text="Photoshop uninstalled."
+echo "Photoshop uninstalled"
